Upload some working files
VanillaWorldBuild.py creates a 2d plot of the IVAN worldmap as it is generated in for ex. IVAN 0.50
WorldBuild_3d.py creates 3d plots of the world map as per IVAN 0.50, with a contour map situated at the floor of the plot. It will generate as output some png images in a subdirectory named "outputs" and will plot at each step. The 3d map is quantized to 4x4 squares
WorldBuild_Fast_3d.py is the same as WorldBuild_3d.py, only it creates only the last plot and png file, but plots the full resolution (1x1 squares)